Fix "toggleFullscreen" and "drawText" API issues

+++ rte/src/api/text.hpp
@@ -30,7 +30,7 @@ namespace BruxAPI {
 
 // Text
 int newFont(int i, int c, int t, bool m, int k); // Doc'd
-void drawText(int f, int x, int y, const std::string& s); // Doc'd
+void drawText(int f, float x, float y, const std::string& s); // Doc'd
 SQInteger chint(HSQUIRRELVM v, int i); // Doc'd
 
 } // namespace BruxAPI
